requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-chess==1.999
gunicorn==21.2.0

